requirements/base.txt
Django==4.2.7
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.1
kafka-python==2.0.2
pika==1.3.2
prometheus-client==0.19.0
django-prometheus==2.3.1
psutil==5.9.6
requests==2.31.0
python-json-logger==2.0.7
django-cors-headers==4.3.1
